+++ ptud_project/Models/ProductModel.cs
@@ -10,7 +10,7 @@ namespace ptud_project.Models
     public class ProductUpdateModel
     {
         public string product_name { get; set; }
-        public Decimal unit_price { get; set; }
+        public Double unit_price { get; set; }
         public string unit_product_name { get; set; }
         public Int64 product_remaining { get; set; }
     }
